refactor: group all user related components and pages together
