+++ src/frontend/components/user-data-components/change-password-components/index.js
@@ -1,13 +1,13 @@
 import { customElement, property } from 'lit/decorators.js';
-import { LitNoShadow } from '../../utils/lit-no-shadow/index.js';
-import { changeUrl } from '../../utils/helpers/change-url.js';
-import { state } from '../../worker/index.js';
+import { LitNoShadow } from '../../../utils/lit-no-shadow/index.js';
+import { changeUrl } from '../../../utils/helpers/change-url.js';
+import { state } from '../../../worker/index.js';
 import { template } from './template.js';
 
 /**
  * @type {LitPage}
  */
-@customElement('register-component')
+@customElement('change-password-component')
 class Component extends LitNoShadow {
   @property({ type: String })
   errorMessage = ''
@@ -15,40 +15,31 @@ class Component extends LitNoShadow {
     return template.bind(this)();
   }
   // this is called when submit button is clicked (see template.js)
-  async register (event) {
+  async changePassword (event) {
     // this prevents the page from using the default behavior
     // of form submit
     event.preventDefault();
     // gets the event.target and change the variable name to form
     const { target: form } = event;
-    console.log(form)
-    const username = form.username.value;
-    const password = form.password.value;
-    const firstName = form.firstName.value;
-    const lastName = form.lastName.value;
-    
+    const newPassword = form.password.value;
     // calls an API call
-    const response = await window.fetch('/api/user', {
+    const response = await window.fetch('/api/change-password', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
       },
       body: JSON.stringify({
-        username,
-        password,
-        firstName,
-        lastName
+        newPassword
       })
     });
     if (response.status === 200) {
       this.errorMessage = '';
-      await state.set('user-is-logged-in', true);
-      return changeUrl('/');
+      return changeUrl('/')
     //   return changeUrl('/todos');
     }
     const { message, error } = await response.json();
     this.errorMessage = `HTTP Code: ${response.status} - ${error} - ${message}`;
-    await state.set('user-is-logged-in', false);
+
   }
 }
