+++ src/frontend/components/user-data-components/header-component/index.js
@@ -1,6 +1,6 @@
 import { customElement, property } from 'lit/decorators.js';
-import { state } from '../../worker/index.js';
-import { LitNoShadow } from '../../utils/lit-no-shadow/index.js';
+import { state } from '../../../worker/index.js';
+import { LitNoShadow } from '../../../utils/lit-no-shadow/index.js';
 import { proxy } from 'comlink';
 import { template } from './template.js';
