+++ src/frontend/components/user-data-components/login-component/index.js
@@ -1,13 +1,13 @@
 import { customElement, property } from 'lit/decorators.js';
-import { LitNoShadow } from '../../utils/lit-no-shadow/index.js';
-import { changeUrl } from '../../utils/helpers/change-url.js';
-import { state } from '../../worker/index.js';
+import { LitNoShadow } from '../../../utils/lit-no-shadow/index.js';
+import { changeUrl } from '../../../utils/helpers/change-url.js';
+import { state } from '../../../worker/index.js';
 import { template } from './template.js';
 
 /**
  * @type {LitPage}
  */
-@customElement('change-password-component')
+@customElement('login-component')
 class Component extends LitNoShadow {
   @property({ type: String })
   errorMessage = ''
@@ -15,31 +15,34 @@ class Component extends LitNoShadow {
     return template.bind(this)();
   }
   // this is called when submit button is clicked (see template.js)
-  async changePassword (event) {
+  async login (event) {
     // this prevents the page from using the default behavior
     // of form submit
     event.preventDefault();
     // gets the event.target and change the variable name to form
     const { target: form } = event;
-    const newPassword = form.password.value;
+    const username = form.username.value;
+    const password = form.password.value;
     // calls an API call
-    const response = await window.fetch('/api/change-password', {
+    const response = await window.fetch('/api/login', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
       },
       body: JSON.stringify({
-        newPassword
+        username,
+        password
       })
     });
     if (response.status === 200) {
       this.errorMessage = '';
+      await state.set('user-is-logged-in', true);
       return changeUrl('/')
     //   return changeUrl('/todos');
     }
     const { message, error } = await response.json();
     this.errorMessage = `HTTP Code: ${response.status} - ${error} - ${message}`;
-
+    await state.set('user-is-logged-in', false);
   }
 }
