+++ src/frontend/components/user-data-components/register-component/template.js
