+++ src/frontend/pages/user-data-pages/page-change-password/index.js
@@ -1,11 +1,11 @@
 import { customElement } from 'lit/decorators.js';
-import { LitPage } from '../../utils/lit-page/index.js';
+import { LitPage } from '../../../utils/lit-page/index.js';
 import { template } from './template.js';
 
 /**
  * @type {LitPage}
  */
-@customElement('page-register')
+@customElement('page-change-password')
 class Page extends LitPage {
   render () {
     return template.bind(this)();
