+++ src/frontend/pages/user-data-pages/page-change-password/template.js
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 // loads the login-component
-import '../../components/change-password-components/index.js';
+import '../../../components/user-data-components/change-password-components/index.js';
 
 export function template () {
   return html`<change-password-component></change-password-component>`;
