+++ src/frontend/pages/user-data-pages/page-change-user-data/template.js
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 // loads the login-component
-import '../../components/change-user-data-component/index.js';
+import '../../components/user-data-components/change-user-data-component/index.js';
 
 export function template () {
   return html`<change-user-data-component></change-user-data-component>`;
