+++ src/frontend/pages/user-data-pages/page-home/index.js
@@ -1,11 +1,11 @@
 import { customElement } from 'lit/decorators.js';
-import { LitPage } from '../../utils/lit-page/index.js';
+import { LitPage } from '../../../utils/lit-page/index.js';
 import { template } from './template.js';
 
 /**
  * @type {LitPage}
  */
-@customElement('page-change-password')
+@customElement('page-home')
 class Page extends LitPage {
   render () {
     return template.bind(this)();
