+++ src/frontend/pages/user-data-pages/page-logout/index.js
@@ -1,7 +1,7 @@
 import { customElement } from 'lit/decorators.js';
-import { state } from '../../worker/index.js';
-import { changeUrl } from '../../utils/helpers/change-url.js';
-import { LitPage } from '../../utils/lit-page/index.js';
+import { state } from '../../../worker/index.js';
+import { changeUrl } from '../../../utils/helpers/change-url.js';
+import { LitPage } from '../../../utils/lit-page/index.js';
 import { template } from './template.js';
 
 /**
