+++ src/frontend/pages/user-data-pages/page-logout/template.js
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 // loads the login-component
-import '../../components/login-component/index.js';
+import '../../../components/user-data-components/login-component/index.js';
 
 export function template () {
   return html`Logging Out`;
