+++ src/frontend/pages/user-data-pages/page-register/index.js
@@ -1,11 +1,11 @@
 import { customElement } from 'lit/decorators.js';
-import { LitPage } from '../../utils/lit-page/index.js';
+import { LitPage } from '../../../utils/lit-page/index.js';
 import { template } from './template.js';
 
 /**
  * @type {LitPage}
  */
-@customElement('page-home')
+@customElement('page-register')
 class Page extends LitPage {
   render () {
     return template.bind(this)();
