+++ src/frontend/pages/user-data-pages/page-register/template.js
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 // loads the login-component
-import '../../components/register-component/index.js';
+import '../../../components/user-data-components/register-component/index.js';
 
 export function template () {
   return html`<register-component></register-component>`;
